indicator-ng: always create entry.label and entry.image
They used to be created lazily when the root menu item changed to include and
image or a label. However, there is no way to notify a panel that this
happened.

Panels connect to notify signals on these widgets when an entry gets added. If
the widgets don't exist from the very beginning, the panel won't update when
they change.