indicator-ng: support "submenu-action" on root menu items
Treat "submenu-action" attributes on the root menu item in the same way that
gtk+ does. If that attribute exists, it must reference an action with boolean
state that is set to 'true' whenever a corresponding indicator menu is open.

We need this for root items because they are not handled by gtk's menu tracker.
Nested submenus already get this.