IndicatorImageHelper: always try to use a GIcon or the file-name as source of the Gdk image
We don't need to fallback to pure pixbuf, unless an indicator provided a custom icon that is
not in the current theme path.
This helps a lot in reducing the Unity7 workload as this decreases the cases where
we need to encode the pixbuf contents, send them via dbus to unity, encode them
back, reload to a new pixbuf...

Also thanks to this, the library clients can load the actual icon, scaled at the value they want.